+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.1.4] - 2025-05-22
+
 ### Added
 
 - Added support to `claude-sonnet-4` and `claude-opus-4` models from Anthropic.
@@ -627,7 +629,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of the `daiv` project.
 
-[Unreleased]: https://github.com/srtab/daiv/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/srtab/daiv/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/srtab/daiv/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/srtab/daiv/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/srtab/daiv/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/srtab/daiv/compare/v0.1.0...v0.1.1